+++ src/__tests__/emitter.spec.ts
@@ -8,26 +8,26 @@ type VT = {
 
 describe("Emitter", () => {
   describe("on", () => {
-    it("should attach a listener function", () => {
+    it("should attach a handler function", () => {
       const emitter = new Emitter<Keys, VT>();
       let x: string | undefined = undefined;
-      const listener = (str: string): void => {
+      const handler = (str: string): void => {
         x = str;
       };
-      emitter.on("str", listener);
+      emitter.on("str", handler);
       emitter.emitSync("str", "foo");
       expect(x).toBe("foo");
       emitter.emitSync("str", "bar");
       expect(x).toBe("bar");
     });
 
-    it("should return a subscription which will remove the listener when unsubscribed", () => {
+    it("should return a subscription which will remove the handler when unsubscribed", () => {
       const emitter = new Emitter<Keys, VT>();
       let x: string | undefined = undefined;
-      const listener = (str: string): void => {
+      const handler = (str: string): void => {
         x = str;
       };
-      const subscription = emitter.on("str", listener);
+      const subscription = emitter.on("str", handler);
       emitter.emitSync("str", "foo");
       expect(x).toBe("foo");
       subscription.unsubscribe();
@@ -39,10 +39,10 @@ describe("Emitter", () => {
       const emitter = new Emitter<Keys, VT>();
       const onStr = emitter.on("str");
       let x: string | undefined = undefined;
-      const listener = (str: string): void => {
+      const handler = (str: string): void => {
         x = str;
       };
-      const subscription = onStr(listener);
+      const subscription = onStr(handler);
       emitter.emitSync("str", "foo");
       expect(x).toBe("foo");
       emitter.emitSync("str", "bar");
@@ -54,34 +54,34 @@ describe("Emitter", () => {
   });
 
   describe("off", () => {
-    it("should remove a listener function", () => {
+    it("should remove a handler function", () => {
       const emitter = new Emitter<Keys, VT>();
       let x: string | undefined = undefined;
-      const listener = (str: string): void => {
+      const handler = (str: string): void => {
         x = str;
       };
-      emitter.on("str", listener);
+      emitter.on("str", handler);
       emitter.emitSync("str", "foo");
       expect(x).toBe("foo");
-      emitter.off("str", listener);
+      emitter.off("str", handler);
       emitter.emitSync("str", "bar");
       expect(x).toBe("foo");
     });
   });
 
   describe("emitSync", () => {
-    it("should synchronously invoke listener functions attached to a key", () => {
+    it("should synchronously invoke handler functions attached to a key", () => {
       const emitter = new Emitter<Keys, VT>();
       let x1: string | undefined = undefined;
-      const listener1 = (str: string): void => {
+      const handler1 = (str: string): void => {
         x1 = str;
       };
       let x2: string | undefined = undefined;
-      const listener2 = (str: string): void => {
+      const handler2 = (str: string): void => {
         x2 = str;
       };
-      emitter.on("str", listener1);
-      emitter.on("str", listener2);
+      emitter.on("str", handler1);
+      emitter.on("str", handler2);
       emitter.emitSync("str", "foo");
       expect(x1).toBe("foo");
       expect(x2).toBe("foo");
@@ -89,18 +89,18 @@ describe("Emitter", () => {
   });
 
   describe("emit", () => {
-    it("should asynchronously invoke listener functions attached to a key", async () => {
+    it("should asynchronously invoke handler functions attached to a key", async () => {
       const emitter = new Emitter<Keys, VT>();
       let x1: string | undefined = undefined;
-      const listener1 = (str: string): void => {
+      const handler1 = (str: string): void => {
         x1 = str;
       };
       let x2: string | undefined = undefined;
-      const listener2 = (str: string): void => {
+      const handler2 = (str: string): void => {
         x2 = str;
       };
-      emitter.on("str", listener1);
-      emitter.on("str", listener2);
+      emitter.on("str", handler1);
+      emitter.on("str", handler2);
       const promise = emitter.emit("str", "foo");
       expect(x1).toBe(undefined);
       expect(x2).toBe(undefined);