+++ src/index.ts
@@ -1,2 +1,2 @@
 export { Unsubscribable, Subscription, CompositeSubscription } from "./subscription";
-export { Listener, Listen, Emitter } from "./emitter";
+export { Handler, Subscribable, Emitter } from "./emitter";